WDSE - Second Evaluation: JPA practice- Validations with Jakarta

+++ Web Development in Server Environment/Second Evaluation/bookstore_jpa/src/main/java/com/fpmislata/bookstore/controller/user/BookUserController.java
@@ -42,7 +42,7 @@ public class BookUserController {
                         .stream()
                         .map(BookMapper.INSTANCE::toBookCollection)
                         .toList(),
-                bookList.getCount(), page, pageSize, baseUrl);
+                (int) bookList.getCount(), page, pageSize, baseUrl);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
